2011-06-18  Evan Martin  <evan@chromium.org>
        Reviewed by Darin Adler.

        Compiler warning in double->int conversion in TestNetscapePlugIn
        https://bugs.webkit.org/show_bug.cgi?id=51554

        This code is providing a JS API for removing a property by numeric ID.
        We get it from JS as a double, and must convert it to an int before using it.

        Test: LayoutTests/plugins/npruntime/remove-property.html

        * DumpRenderTree/TestNetscapePlugIn/Tests/NPRuntimeRemoveProperty.cpp:
        (NPRuntimeRemoveProperty::TestObject::invoke):

Canonical link: https://commits.webkit.org/78531@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@89202 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/DumpRenderTree/TestNetscapePlugIn/Tests/NPRuntimeRemoveProperty.cpp
@@ -63,7 +63,7 @@ private:
             
                 propertyName = pluginTest()->NPN_GetStringIdentifier(propertyNameString.c_str());
             } else {
-                int32_t number = arguments[1].value.doubleValue;
+                int32_t number = static_cast<int32_t>(arguments[1].value.doubleValue);
                 propertyName = pluginTest()->NPN_GetIntIdentifier(number);
             }
